[AMAX][ZenHeart][SmileMiiror][TT-986185]在對談模式中，出現微笑偵測框
root cause:
activity is killed unexpectedly to free up memory when the app idle
in the background for a long time. ViewPager position is restored to
where is was before through bundle but not does the member field(mode state)
when the activity is restarted.

solution: there are two ways to solve:
a.save mode state with onSaveInstanceState and restore with onRestoreInstanceState
b.reset Viewpager to the initial state(first page) with setCurrentItem(0)

(b) is chose because this activity is orientation-locked,which means rotation is not to be handled
each state is expected to be "re"-initial when the activity is "re"-started.

Change-Id: Ifa3d7302a1726f02ac743314187ecbf3e559e1e3
Reviewed-on: http://amax01:8080/368609
Reviewed-by: Jack Tsai <Jack_Tsai@asus.com>
Tested-by: Jack Tsai <Jack_Tsai@asus.com>
Reviewed-by: Feiwen Cheng <feiwen_cheng@asus.com>
Tested-by: Feiwen Cheng <feiwen_cheng@asus.com>